% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jordan_pts.R
\name{get_jordan_pts}
\alias{get_jordan_pts}
\title{Michael Jordan's Points}
\usage{
get_jordan_pts(df, year = 1986)
}
\arguments{
\item{df}{The data you will be using for this function.}

\item{Year}{what year do you want? Defaults to 1986}
}
\description{
This function provides you MJ's total points for a given year
}
\examples{
get_jordan_pts()
}
\keyword{,}
\keyword{mj}
\keyword{pts}
