% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_correlation.R
\name{show_correlation}
\alias{show_correlation}
\title{NBA Correlation}
\usage{
show_correlation(df, year = 1986)
}
\arguments{
\item{df}{The data you will be using}

\item{Year}{what year do you want? Defaults to 1986}
}
\description{
This function provides you with the correlation between statistics for a given year.
}
\examples{
show_correlation()
}
\keyword{,}
\keyword{NBA}
\keyword{correlation}
